feat: add extractFeatures helper
