+++ packages/tfjs-node-helpers/src/feature-engineering/prepare-datasets-for-binary-classification.ts
@@ -1,6 +1,7 @@
 import { data, TensorContainer } from '@tensorflow/tfjs-node';
 import { splitSamplesIntoTrainingValidationTestForBinaryClassification } from '../data-splitting/training-validation-test-for-binary-classification';
 import { makeDataset } from '../utils/make-dataset';
+import { extractFeatures } from './extract-features';
 import { FeatureExtractor } from './feature-extractor';
 
 export const prepareDatasetsForBinaryClassification = async <D, T>({
@@ -24,23 +25,11 @@ export const prepareDatasetsForBinaryClassification = async <D, T>({
   validationDataset: data.Dataset<TensorContainer>;
   testingDataset: data.Dataset<TensorContainer>;
 }> => {
-  const samples = [];
-
-  for (const dataItem of data) {
-    const [inputFeatures, outputFeature] = await Promise.all([
-      Promise.all(
-        inputFeatureExtractors.map((featureExtractor) => {
-          return featureExtractor.extract(dataItem);
-        })
-      ),
-      outputFeatureExtractor.extract(dataItem)
-    ]);
-
-    const input = inputFeatures.map((feature) => feature.value);
-    const output = [outputFeature.value];
-
-    samples.push({ input, output });
-  }
+  const samples = await extractFeatures({
+    data,
+    inputFeatureExtractors,
+    outputFeatureExtractor
+  });
 
   const { trainingSamples, validationSamples, testingSamples } = splitSamplesIntoTrainingValidationTestForBinaryClassification(
     samples,
