+++ packages/tfjs-node-helpers/src/index.ts
@@ -1,6 +1,7 @@
 export * from './classification/binary-classification-trainer';
 export * from './classification/binary-classifier';
 export * from './data-splitting/training-validation-test-for-binary-classification';
+export * from './feature-engineering/extract-features';
 export * from './feature-engineering/feature-extractor';
 export * from './feature-engineering/feature';
 export * from './feature-engineering/prepare-datasets-for-binary-classification';
